requirements.txt
requests
numpy
python-decouple
telebot
python-telegram-bot
pyyaml